backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymysql==1.1.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
razorpay==1.4.1
reportlab==4.0.9
qrcode==7.4.2
Pillow==10.2.0
python-dotenv==1.0.1
httpx==0.26.0
email-validator
